rename slides fixtures in player
